Update JSON loading and testing
Rather than explicitly referencing a URL in code, we define a YAML file
with the specified endpoints, and use Rails' configuration helpers to
make this available system-wide, e.g. for the `articles` endpoint:

    Rails.configuration.json_path.articles

or

    Rails.configuration.json_path[:articles]

This allows per-environment overrides. We provide an (invalid) test URL
for the test environment so that live data can never be referenced.

We also create a test helper that wraps `net/http`. It can accept a
symbol referencing the configuration file (e.g., `:articles`) or a full
URL.

When testing we can explicitly override the URL in the test config.
As we are now doing this in two places in the test code, the act of
mocking a remote URL to return custom JSON is extracted into a helper
method.