Get all likes in a single query
By using a `GROUP BY` query we can get a count of each article's likes
as a hash, with the article ID as the key and the number of likes as the
value. By updating the hash that ActiveRecord returns and setting its
default value to be `0`, we can avoid checking for nil values in the
view layer.