Make each partial independent of the others
Passing the entire hash of likes to each partial is inefficient and
doesn't allow us to keep each partial separate from the others. We can
optimise things by passing in the number of likes to each partial.
This means that we have to handle the looping ourselves instead of
passing the collection to a single `render` call.