Allow system specs to use multiple expectations
When modelling a user's route through a particular feature, we generally
check various expectations at different points of their journey. This
makes the general good rule of "one expectation per spec" too
proscriptive, so we can relax that rule for system tests only.

This is a better approach than using `:aggregate_failures` on all
tests. When a system spec fails, we want it to fail quickly; it is
more than likely subsequent expectations will also fail.

On the other hand, if checking multiple elements of a page, it can
still be worth wrapping those in an `aggregate_failures` block or
tagging the spec, so that all the expectations are checked at once.