Always include JSON fixture helpers in system specs
As we always want the JSON helpers available when writing system specs,
set up an RSpec initializer to do that work for us. The trade-off is
that it makes it a little less obvious where mixed in methods such as
`mock_json_response` reside.

I've opted not to set up the mocking of the products endpoint, even
though it is used nearly everywhere in the current system specs. Doing
so would make it harder to test different scenarios in future, should
we want to add other dummy JSON responses.